tweaking the server files a bit
