+++ stephen_schroeder/client_server.js
@@ -1,16 +1,6 @@
 const net = require('net');
 
-const sockets = [];
+const client = net.connect(3000, () => {
 
-net.createServer((socket) => {
-  sockets.push(socket);
-  socket.on('data', (chunk) => {
-    console.log(chunk.toString());
-    socket.write('MESSAGE RECEIVED');
-    sockets.forEach(s) => {
-      s.write('BROADCASTING: ' + chunk.toString())
-    });
-  });
-}).listen(3000, () => {
-  console.log('LISTENING ON PORT 3000');
+  client.write('MESSAGE RECEIVED');
 });
